requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2.12
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-ipware==2.1.0
django-timezone-field==4.0
djangorestframework==3.11.0
dnspython==1.16.0
eventlet==0.25.2
greenlet==0.4.15
httpie==2.1.0
idna==2.9
importlib-metadata==1.6.0
Jinja2==2.11.2
kombu==4.6.8
MarkupSafe==1.1.1
monotonic==1.5
numpy==1.18.4
Pygments==2.6.1
python-crontab==2.4.2
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
